Naprawienie usuwania elementów z bazy danych

+++ backend/src/main/java/com/example/backend/service/DeleteShoppingList.java
@@ -2,6 +2,7 @@ package com.example.backend.service;
 
 import com.example.backend.model.Product;
 import com.example.backend.model.ProductsInLists;
+import com.example.backend.model.ShoppingList;
 import com.example.backend.repositories.ProductsInListsRepository;
 import com.example.backend.repositories.ShoppingListRepository;
 
@@ -10,20 +11,25 @@ import java.util.List;
 
 public class DeleteShoppingList {
 
-    public static void deleteShoppingList(Long shoppingListId, List<ProductsInLists> actProdInLists, ProductsInListsRepository productsInListsRepository, ShoppingListRepository shoppingListRepository) {
-        Iterator<ProductsInLists> it = actProdInLists.iterator();
+    public static void deleteShoppingList(ShoppingList sl, ProductsInListsRepository productsInListsRepository, ShoppingListRepository shoppingListRepository) {
+        Long shoppingListId = sl.getId();
+        Iterator<ProductsInLists> it = sl.getProductsInLists().iterator();
         while(it.hasNext())
         {
             //remove productsInLists in Product
             ProductsInLists pInL = it.next();
             Product p = pInL.getProduct();
-            p.getProductsInListsList().remove(pInL);
+            p.getProductsInLists().remove(pInL);
             //remove products in lists in ShoppingList
-            actProdInLists.remove(pInL);
+            //sl.getProductsInLists().remove(pInL); <- FORBIDDEN USAGE
+            it.remove(); //<- PROPER USAGE
             //delete ProductsInLists
             productsInListsRepository.deleteById(pInL.getId());
         }
         //delete shopping list in ShoppingList
-        shoppingListRepository.deleteById(shoppingListId);
+        if(shoppingListRepository.findById(shoppingListId).isPresent())
+        {
+            shoppingListRepository.deleteById(shoppingListId);
+        }
     }
 }
